feat: BookMongoRepository should be independent to only delete book using user id and book id, not needding find user id inside your logic

+++ src/data/usecases/book/db.delete.book.ts
@@ -1,16 +1,16 @@
 import { DeleteBook } from "@/domain/usecases"
-import { CheckAccountByAccessTokenRepository, DeleteBookRepository } from "@/data/protocols"
+import { DeleteBookRepository, LoadAccountByTokenRepository } from "@/data/protocols"
 
 export class DbDeleteBook implements DeleteBook {
     constructor (
-        private readonly checkAccountByAccessTokenRepository: CheckAccountByAccessTokenRepository,
+        private readonly loadAccountByTokenRepository: LoadAccountByTokenRepository,
         private readonly deleteBookRepository: DeleteBookRepository,
     ) {}
 
     async delete (data: DeleteBook.Params): Promise<boolean> {
-        const hasAccount = await this.checkAccountByAccessTokenRepository.checkByAccessToken(data.accessToken)
-        if (hasAccount) {
-            await this.deleteBookRepository.delete(data)
+        const account = await this.loadAccountByTokenRepository.loadByToken(data.accessToken)
+        if (account) {
+            await this.deleteBookRepository.delete({ userId: account.id, bookId: data.bookId })
             return true
         }
         return false
